refactor: UI 함수 폴더 구조 변경
