+++ src/handlers/handlePurchase.js
@@ -1,12 +1,12 @@
-import createLottoInput from './View/create/createLottoInput.js';
-import showLottoResult from './View/show/showLottoResult.js';
-import showPurchaseResult from './View/show/showPurchaseResult.js';
-import SELECTORS from './constants/Selectors.js';
-import { getUIPurchasePrice } from './service/InputService/UIInputService.js';
-import makeLotto from './service/LottoService.js';
-import { getPurchasePrice } from './service/ParsingService.js';
-import state from './state.js';
-import { handlePriceError } from './util/errorHandler.js';
+import createLottoInput from '../View/create/createLottoInput.js';
+import showLottoResult from '../View/show/showLottoResult.js';
+import showPurchaseResult from '../View/show/showPurchaseResult.js';
+import SELECTORS from '../constants/Selectors.js';
+import { getUIPurchasePrice } from '../service/InputService/UIInputService.js';
+import makeLotto from '../service/LottoService.js';
+import { getPurchasePrice } from '../service/ParsingService.js';
+import state from '../state.js';
+import { handlePriceError } from '../util/errorHandler.js';
 
 async function handlePurchase() {
   const purchaseButton = document.getElementById(SELECTORS.BUTTON.PURCHASE);
